Update first_simple_+_code.py
change int to float and deleted empty lines

+++ first_simple_+_code.py
@@ -1,5 +1,7 @@
 a = 5
-b = int(input("pls,write a number:"))
+#change float
+#b = int(input("pls,write a number:"))
+b = float(input("pls,write a number:"))
 c = a+b
 print(c)
-
+#deleted empty lines
